pyproject.toml
[tool.poetry]
name = "disquest-utils"
version = "1.0.5"
description = "A set of async utils for Miku's DisQuest Cog"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/No767/DisQuest-Utils"
repository = "https://github.com/No767/DisQuest-Utils"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
  "Intended Audience :: Developers",
]
include = ["license"]
packages = [{ include = "disquest_utils", from = "src" }]
[tool.poetry.dependencies]
python = "^3.7,<4.0"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.36"
uvloop = "^0.16.0"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
